Весь код по резервуарным вычислениям теперь в одной папке: сам скрипт с функциями, скрипт запуска на кластере оптимизации, файл со стилем для Matplotlib. Оптимизация нелинейного шага в SSFM при наличии Esat.

+++ fiberprop/matrices.py
@@ -40,7 +40,7 @@ def create_simple_dispersion_free_matrix(mat, alpha, g_0, step):
 def create_freq_matrix(mat, beta1, beta2, alpha, g_0, omega, step):
     n = len(mat)
     m = len(omega)
-    asw = np.empty((n * n, m), dtype=np.complex128)
+    asw = np.empty((n * n, m), dtype=mat.dtype)
 
     diag_elements = step * (1j * np.diag(mat)[:, None] + (-2j * beta1[:, None] * omega + 1j * beta2[:, None] * omega ** 2 - alpha[:, None] - g_0[:, None]) / 2)
     off_diag_elements = step * 1j * mat
